+++ types/beatmap/v3/heck.ts
@@ -32,8 +32,7 @@ export interface IHeckCustomEventDataAnimateTrack extends IHeckCustomEventDataBa
 /** AssignPathAnimation interface for Heck Custom Event.
  * @extends IHeckCustomEventDataBase
  */
-export interface IHeckCustomEventDataAssignPathAnimation
-    extends IHeckCustomEventDataBase {
+export interface IHeckCustomEventDataAssignPathAnimation extends IHeckCustomEventDataBase {
     duration: number;
     easing?: Easings;
     position?: string | Vector3PointDefinition[];